requirements.txt
Flask==3.0.3
gunicorn==22.0.0
gspread==6.1.2
google-auth==2.32.0
